requirements.txt
discord.py
python-dotenv
yt-dlp
PyNaCl
Flask
gTTS